[DevTools] Add resizable sidebar to Profiler tab and toolbar compaction
Add drag-to-resize columns to the Profiler panel — the sidebar was
previously fixed at max 300px and is now user-resizable with persisted
state, matching the Components tab behavior. The existing resize logic
is extracted into a shared useResizableColumns hook as part of this.

Add progressive toolbar compaction for the Profiler — when the left
column becomes too narrow, tab labels collapse to icons only, then
the root selector and snapshot selector wrap to a second row.
