+++ packages/react-devtools-shared/src/devtools/views/Components/Components.css
@@ -45,6 +45,7 @@
   cursor: ew-resize;
 }
 
+/* Must match VERTICAL_MODE_MAX_WIDTH in useResizableColumns.js */
 @container devtools (width < 600px) {
   .Components {
     flex-direction: column;
