+++ packages/react-devtools-shared/src/devtools/views/Profiler/CommitRanked.css
@@ -2,4 +2,5 @@
   width: 100%;
   flex: 1;
   padding: 0.5rem;
+  overflow: hidden;
 }
