+++ packages/react-devtools-shared/src/devtools/views/Profiler/Profiler.js
@@ -8,7 +8,14 @@
  */
 
 import * as React from 'react';
-import {Fragment, useContext, useEffect, useRef, useEffectEvent} from 'react';
+import {
+  useContext,
+  useEffect,
+  useLayoutEffect,
+  useRef,
+  useState,
+  useEffectEvent,
+} from 'react';
 import {ModalDialog} from '../ModalDialog';
 import {ProfilerContext} from './ProfilerContext';
 import TabBar from '../TabBar';
@@ -32,13 +39,32 @@ import SettingsModal from 'react-devtools-shared/src/devtools/views/Settings/Set
 import SettingsModalContextToggle from 'react-devtools-shared/src/devtools/views/Settings/SettingsModalContextToggle';
 import {SettingsModalContextController} from 'react-devtools-shared/src/devtools/views/Settings/SettingsModalContext';
 import portaledContent from '../portaledContent';
+import useResizableColumns from '../useResizableColumns';
 import {StoreContext} from '../context';
 import {TimelineContext} from 'react-devtools-timeline/src/TimelineContext';
 
 import styles from './Profiler.css';
 
+const LOCAL_STORAGE_KEY = 'React::DevTools::profilerResizeState';
+
+// Toolbar compaction levels:
+// 0 = full labels, everything on one row
+// 1 = tab icons only, everything on one row
+// 2 = tab icons only, root selector + snapshot on second row
+const COMPACT_NONE = 0;
+const COMPACT_ICONS = 1;
+const COMPACT_WRAP = 2;
+
 function Profiler(_: {}) {
-  const profilerRef = useRef<HTMLDivElement | null>(null);
+  const {wrapperRef, resizeElementRef, onResizeStart, onResizeEnd, onResize} =
+    useResizableColumns(LOCAL_STORAGE_KEY);
+  const toolbarRef = useRef<HTMLDivElement | null>(null);
+  const [compactLevel, setCompactLevel] = useState(COMPACT_NONE);
+  // Remembers how wide the toolbar content was when it overflowed.
+  // We only expand back when the container grows past that width,
+  // so we don't get stuck in a compact → expand → overflow → compact loop.
+  const expandThresholdsRef = useRef<Array<number>>([0, 0]);
+
   const isMac =
     typeof navigator !== 'undefined' &&
     navigator.platform.toUpperCase().indexOf('MAC') >= 0;
@@ -98,7 +124,7 @@ function Profiler(_: {}) {
   });
 
   useEffect(() => {
-    const div = profilerRef.current;
+    const div = wrapperRef.current;
     if (!div) {
       return;
     }
@@ -108,6 +134,40 @@ function Profiler(_: {}) {
       ownerWindow.removeEventListener('keydown', handleKeyDown);
     };
   }, []);
+
+  useLayoutEffect(() => {
+    const toolbar = toolbarRef.current;
+    if (toolbar == null) {
+      return;
+    }
+
+    const checkOverflow = () => {
+      if (toolbar.scrollWidth > toolbar.clientWidth) {
+        // Content doesn't fit — remember how wide it was, then compact more.
+        if (compactLevel < COMPACT_WRAP) {
+          expandThresholdsRef.current[compactLevel] = toolbar.scrollWidth;
+          setCompactLevel(compactLevel + 1);
+        }
+      } else if (compactLevel > COMPACT_NONE) {
+        // Content fits — but only expand back if the container is now
+        // at least as wide as the content was before we compacted.
+        const prevThreshold = expandThresholdsRef.current[compactLevel - 1];
+        if (prevThreshold > 0 && toolbar.clientWidth >= prevThreshold) {
+          setCompactLevel(compactLevel - 1);
+        }
+      }
+    };
+
+    const observer = new ResizeObserver(checkOverflow);
+    observer.observe(toolbar);
+
+    // Check immediately for content changes (e.g. profile data loaded)
+    checkOverflow();
+
+    return () => observer.disconnect();
+    // Re-check when didRecordCommits changes because new elements
+    // (e.g. SnapshotSelector) appear and may cause overflow.
+  }, [compactLevel, didRecordCommits]);
 
   let view = null;
   if (didRecordCommits || selectedTabID === 'timeline') {
@@ -165,22 +225,23 @@ function Profiler(_: {}) {
 
   return (
     <SettingsModalContextController>
-      <div ref={profilerRef} className={styles.Profiler}>
-        <div className={styles.LeftColumn}>
-          <div className={styles.Toolbar}>
+      <div ref={wrapperRef} className={styles.Profiler}>
+        <div ref={resizeElementRef} className={styles.LeftColumn}>
+          <div ref={toolbarRef} className={styles.Toolbar}>
             <RecordToggle disabled={!supportsProfiling} />
             <ReloadAndProfileButton disabled={!supportsProfiling} />
             <ClearProfilingDataButton />
             <ProfilingImportExportButtons />
             <div className={styles.VRule} />
             <TabBar
+              compact={compactLevel >= COMPACT_ICONS}
               currentTab={selectedTabID}
               id="Profiler"
               selectTab={selectTab}
               tabs={supportsTimeline ? tabsWithTimeline : tabs}
               type="profiler"
             />
-            <RootSelector />
+            {compactLevel < COMPACT_WRAP && <RootSelector />}
             <div className={styles.Spacer} />
             {!isLegacyProfilerSelected && (
               <div
@@ -189,17 +250,30 @@ function Profiler(_: {}) {
               />
             )}
             <SettingsModalContextToggle />
-            {isLegacyProfilerSelected && didRecordCommits && (
-              <Fragment>
-                <div className={styles.VRule} />
-                <SnapshotSelector />
-              </Fragment>
-            )}
+            {isLegacyProfilerSelected &&
+              didRecordCommits &&
+              compactLevel < COMPACT_WRAP && <SnapshotSelector />}
           </div>
+          {compactLevel >= COMPACT_WRAP && (
+            <div className={styles.SnapshotSelectorWrapper}>
+              <RootSelector />
+              {isLegacyProfilerSelected && didRecordCommits && (
+                <SnapshotSelector />
+              )}
+            </div>
+          )}
           <div className={styles.Content}>
             {view}
             <ModalDialog />
           </div>
+        </div>
+        <div className={styles.ResizeBarWrapper}>
+          <div
+            onPointerDown={onResizeStart}
+            onPointerMove={onResize}
+            onPointerUp={onResizeEnd}
+            className={styles.ResizeBar}
+          />
         </div>
         <div className={styles.RightColumn}>{sidebar}</div>
         <SettingsModal />
